requirements.txt
anaconda-client==1.12.0
anaconda-navigator==2.4.2
ansi2html==1.8.0
appdirs==1.4.4
atomicwrites==1.4.0
brotlipy==0.7.0
click-plugins==1.1.1
cligj==0.7.2
clyent==1.2.2
conda-repo-cli==1.0.41
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
daal4py==2023.1.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datashape==0.5.4
et-xmlfile==1.1.0
Fiona==1.9.4.post1
fonttools==4.25.0
geopandas==0.13.2
gunicorn==21.2.0
inflection==0.5.1
jsonpointer==2.1
mkl-fft==1.3.6
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.4.0
openpyxl==3.0.10